Fixed  teacher, employee, academic to working condition

+++ internal/model/subject.go
@@ -17,14 +17,14 @@ const (
 )
 
 type Subject struct {
-	SubjectID   int64  `db:"subject_id" json:"subject_id"`
+	SubjectID   uint64 `db:"subject_id" json:"subject_id"`
 	Name        string `db:"name" json:"name"`
 	Semester    []int8 `db:"semester" json:"semester"`
 	SubjectType string `db:"subject_type" json:"subject_type"`
 }
 
 type Lesson struct {
-	LessonID    int64     `db:"lesson_id" json:"lesson_id"`
+	LessonID    uint64    `db:"lesson_id" json:"lesson_id"`
 	LectureRoom string    `db:"lecture_room" json:"lecture_room"`
 	Date        time.Time `db:"date" json:"date"`
 	LessonType  string    `db:"lesson_type" json:"lesson_type"`
@@ -38,14 +38,14 @@ type CreateSubjectInput struct {
 	SubjectType string `db:"subject_type" json:"subject_type"`
 }
 type UpdateSubjectInput struct {
-	Name        string  `db:"name" json:"name"`
-	Semester    []uint8 `db:"semester" json:"semester"`
-	SubjectType string  `db:"subject_type" json:"subject_type"`
+	Name        string   `db:"name" json:"name"`
+	Semester    []uint16 `db:"semester" json:"semester"`
+	SubjectType string   `db:"subject_type" json:"subject_type"`
 }
 
 type CreateLessonInput struct {
-	TeacherID   int64     `db:"teacher_id" json:"teacher_id"`
-	SubjectID   int64     `db:"subject_id" json:"subject_id"`
+	TeacherID   uint64    `db:"teacher_id" json:"teacher_id"`
+	SubjectID   uint64    `db:"subject_id" json:"subject_id"`
 	LectureRoom string    `db:"lecture_room" json:"lecture_room"`
 	Date        time.Time `db:"date" json:"date"`
 	LessonType  string    `db:"lesson_type" json:"lesson_type"`
@@ -53,12 +53,12 @@ type CreateLessonInput struct {
 
 type AttendanceRecord struct {
 	StudentBriefInfo
-	LessonID int64  `db:"lesson_id" json:"lesson_id"`
+	LessonID uint64 `db:"lesson_id" json:"lesson_id"`
 	Status   string `db:"status" json:"status"`
 }
 
 type Grade struct {
 	StudentBriefInfo
-	LessonID int64 `db:"lesson_id" json:"lesson_id"`
-	Grade    uint8 `db:"status" json:"status"`
+	LessonID uint64 `db:"lesson_id" json:"lesson_id"`
+	Grade    uint16 `db:"status" json:"status"`
 }
